remove my TODOs, FIXME and some warnings
There is still something to test: an old
client (pre-several ships) vs. new client
(as a server and as client).
Old as a server, the new client should
not be able to play with several ships.
Old as client, the server should not
be able to play with several ships.

+++ src/ai/smartai.cpp
@@ -351,7 +351,7 @@ Strategy* SmartAI::State::defaultStrategy(Sea::Player player, Sea* sea)
 
 void SmartAI::State::destroyed(int size)
 {
-    if (size <= m_config->longestShip()) {
+    if ( size <= static_cast<int>(m_config->longestShip()) ) {
         int index = size - 1;
         if (m_ships[index] > 0) {
             m_ships[index]--;
